Better code
